Start

+++ src/pages/admin/users/index.jsx
@@ -8,6 +8,7 @@ import {
   Input,
   Modal,
   Pagination,
+  Select,
   Space,
   Table,
   Upload,
@@ -29,17 +30,30 @@ import TextArea from "antd/es/input/TextArea";
 const UsersPage = () => {
   const [search, setSearch] = useState("");
   const [page, setPage] = useState(1);
+  const [isRole, setIsRole] = useState("all");
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isModalLoading, setIsModalLoading] = useState(false);
   const [selected, setSelected] = useState(null);
   const [photo, setPhoto] = useState(null);
   const [loadingPhoto, setLoadingPhoto] = useState(false);
   const [form] = Form.useForm();
+  let params = {
+    page,
+    search,
+  };
+  if (isRole !== "all") {
+    params = {
+      page,
+      search,
+      role: isRole,
+    };
+  }
+
   const {
     data: { users, total } = { users: [], total: 0 },
     isFetching,
     refetch,
-  } = useGetUsersQuery({ page, search });
+  } = useGetUsersQuery({ ...params });
   const [uploadPhoto] = useUploadPhotoMutation();
   const [addUser] = useAddUserMutation();
   const [editUser] = useEditUserMutation();
@@ -51,12 +65,20 @@ const UsersPage = () => {
     setPage(1);
   };
 
+  const handleChange = (value) => {
+    setIsRole(value);
+  };
+
   const showModal = () => {
     setIsModalOpen(true);
+    form.resetFields();
+    setSelected(null);
+    setPhoto(null);
   };
 
   const closeModal = () => {
     setIsModalOpen(false);
+    form.resetFields();
   };
 
   const handlePhoto = async (e) => {
@@ -64,9 +86,8 @@ const UsersPage = () => {
     const formData = new FormData();
     formData.append("file", e.file.originFileObj);
     const { data } = await uploadPhoto(formData);
-    setPhoto(data);
     setLoadingPhoto(false);
-    console.log(data);
+    setPhoto(data);
   };
 
   const handleOk = async () => {
@@ -75,8 +96,7 @@ const UsersPage = () => {
       const values = await form.validateFields();
       values.photo = photo;
       if (selected === null) {
-        const { data } = await addUser(values);
-        console.log(data);
+        await addUser(values);
       } else {
         await editUser({ id: selected, body: values });
       }
@@ -92,7 +112,7 @@ const UsersPage = () => {
   const handleEdit = async (id) => {
     setSelected(id);
     const { data } = await getUser(id);
-    setPhoto(data?.photo)
+    setPhoto(data?.photo);
     setIsModalOpen(true);
     form.setFieldsValue(data);
   };
@@ -115,7 +135,34 @@ const UsersPage = () => {
           style={{ width: "auto", flexGrow: 1 }}
           placeholder="Searching..."
         />
-        <Button onClick={showModal} type="dashed">
+        <Select
+          value={isRole}
+          onChange={(value) => {
+            handleChange(value);
+          }}
+          style={{
+            width: 120,
+          }}
+          options={[
+            {
+              value: "all",
+              label: "All",
+            },
+            {
+              value: "admin",
+              label: "Admin",
+            },
+            {
+              value: "client",
+              label: "Client",
+            },
+            {
+              value: "user",
+              label: "User",
+            },
+          ]}
+        />
+        <Button onClick={showModal} type="primary">
           Add users
         </Button>
       </Flex>
@@ -173,7 +220,12 @@ const UsersPage = () => {
                   danger
                   type="primary"
                   onClick={() => {
-                    handleDelete(data?._id);
+                    Modal.confirm({
+                      title: "Do you want to delete this user?",
+                      onOk: () => {
+                        handleDelete(data?._id);
+                      },
+                    });
                   }}
                 >
                   Delete
